refactor: split audio and storage into modules
