+++ icy-tower/index.html
@@ -123,6 +123,6 @@
     <button id="newGameBtn">Nowa Gra</button>
   </div>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/tone/14.8.39/Tone.min.js"></script>
-  <script src="game.js"></script>
+  <script type="module" src="src/core.js"></script>
 </body>
 </html>
